Do a full nullability spec across the Pod
I went through the code for the Pod and specified nullability
in all the header and implementation files. This is done to
improve interoperability with Swift's optional types, as Swift
currently assumes all values returned by the API are non-null,
while the Pod uses nil values pretty frequently to convey state.